✏️ Fix: rename movie error code and review message file
