+++ src/app/board/reviews/page.tsx
@@ -1,7 +1,7 @@
 import { auth } from '@/auth'
 import GoBackHeader from '@/components/layout/MobileHeader/GoBackHeader'
 import ReviewList from '@/components/ui/ReviewList'
-import { MOVIES_CODES } from '@/constants/messages/movies'
+import { MOVIES_CODES } from '@/constants/messages/movie'
 import { getThisWeekMovieId } from '@/lib/api/discussion'
 import { getReviews } from '@/lib/api/review'
 import { ReviewSortField } from '@/types/api/common'
@@ -31,7 +31,7 @@ export default async function BoardReviewsPage({
     })
   } catch (error) {
     const errorCode = (error as Error).message
-    if (errorCode === MOVIES_CODES.MOVIES_NOT_FOUND) return notFound()
+    if (errorCode === MOVIES_CODES.MOVIE_NOT_FOUND) return notFound()
     throw error
   }
