+++ src/app/reviews/[id]/CommentForm.tsx
@@ -9,7 +9,7 @@ import { ClientComment, Comment } from '@/types/api/common'
 import { useMobileKeyboard } from '@/hooks/useMobileKeyboard'
 import { COMMON_CODES, COMMON_MESSAGES } from '@/constants/messages/common'
 import toast from 'react-hot-toast'
-import { REVIEW_CODES, REVIEW_MESSAGES } from '@/constants/messages/reviews'
+import { REVIEW_CODES, REVIEW_MESSAGES } from '@/constants/messages/review'
 import { useRouter } from 'next/navigation'
 import { useForm } from 'react-hook-form'
 import { CommentsRequest } from '@/types/api/comment'
