+++ src/components/form/ReviewForm.tsx
@@ -7,7 +7,7 @@ import { useRouter } from 'next/navigation'
 import CertifiedBadge from '../ui/CertifiedBadge'
 import { COMMON_CODES, COMMON_MESSAGES } from '@/constants/messages/common'
 import toast from 'react-hot-toast'
-import { REVIEW_CODES, REVIEW_MESSAGES } from '@/constants/messages/reviews'
+import { REVIEW_CODES, REVIEW_MESSAGES } from '@/constants/messages/review'
 import { useForm } from 'react-hook-form'
 import { baseReviewClientInput, baseReviewClientSchema } from '@/schemas/common/baseReview'
 import { zodResolver } from '@hookform/resolvers/zod'
