+++ src/lib/actions/like.ts
@@ -3,7 +3,7 @@
 import { auth } from '@/auth'
 import { toggleLike } from '../api/review'
 import { COMMON_CODES } from '@/constants/messages/common'
-import { REVIEW_CODES } from '@/constants/messages/reviews'
+import { REVIEW_CODES } from '@/constants/messages/review'
 import { redirect } from 'next/navigation'
 
 export async function toggleLikeAction(
